+++ src/operators.ts
@@ -1,13 +1,17 @@
-export default [
-  { key: "EQUALS", label: "equals", type: "STRING" },
-  { key: "CONTAINS", label: "contains", type: "STRING" },
-  { key: "NOT_CONTAIN", label: "not contains", type: "STRING" },
-  { key: "NOT_EQUAL", label: "is empty", type: "STRING" },
-  { key: "IS_EMPTY", label: "is empty", type: "STRING" },
-
-  { key: "EQUALS", label: "equals", type: "NUMBER" },
-  { key: "LESS_THAN", label: "less than", type: "NUMBER" },
-  { key: "GREATER_THAN", label: "greater than", type: "NUMBER" },
-
-  { key: "EQUALS", label: "not empty", type: "BOOLEAN" }
-];
+export default {
+  string: [
+    { key: "EQUALS", label: "equals", type: "STRING" },
+    { key: "CONTAINS", label: "contains", type: "STRING" },
+    { key: "NOT_CONTAIN", label: "not contains", type: "STRING" },
+    { key: "NOT_EQUAL", label: "is empty", type: "STRING" },
+    { key: "IS_EMPTY", label: "is empty", type: "STRING" }
+  ],
+  number: [
+    { key: "EQUALS", label: "equals", type: "NUMBER" },
+    { key: "LESS_THAN", label: "less than", type: "NUMBER" },
+    { key: "GREATER_THAN", label: "greater than", type: "NUMBER" }
+  ],
+  boolean: [
+    { key: "EQUALS", label: "equals", type: "BOOLEAN" }
+  ]
+};
