+++ src/partials/query.editor.html
@@ -197,7 +197,7 @@
       <div ng-repeat="filter in ctrl.target.filters">
         <div class="gf-form">
           <label class="gf-form-label width-13 pointer" for="in-filters-{{ctrl.target.refId}}-{{$index}}">
-            {{$index+1}}.
+            {{$index+1}}. Filter
             <info-popover position="top center" class="ng-invalid" mode="right-normal">
               <p>Filter by tag</p>
             </info-popover>
@@ -208,6 +208,7 @@
                     ng-model="ctrl.target.filters[$index].tag"
                     ng-change="ctrl.onTagFilterChange($index)"
                     ng-options="f as f.key + ' (' + f.type + ')' for f in ctrl.uniqueTags track by f.key">
+              <option value="" disabled selected>Pick a Tag</option>
             </select>
           </div>
           <div class="gf-form-select-wrapper width-8 gf-form-select-wrapper--has-help-icon">
@@ -215,7 +216,7 @@
                     class="gf-form-input ng-pristine ng-valid ng-not-empty ng-touched"
                     ng-model="ctrl.target.filters[$index].operator"
                     ng-change="ctrl.onTagFilterChange($index)"
-                    ng-options="f as f.label for f in ctrl.uniqueOperators track by f.key">
+                    ng-options="f as f.label for f in ctrl.uniqueOperators['string'] track by f.key">
             </select>
           </div>
           <input id="in-filter-stringValue-{{ctrl.target.refId}}-{{$index}}"
@@ -226,7 +227,6 @@
                  ng-change="ctrl.onTagFilterChange($index)"
                  placeholder="Please Specify"
                  spellcheck='false'/>
-          {{ctrl.target.filters[$index].operator.type}}
           <button class="btn btn-inverse" ng-click="ctrl.removeFilter($index)">
   			    <i class="fa fa-minus"></i>
   		    </button>
